+++ app/App/Transactions/TransactionDetailsModal.js
@@ -40,8 +40,9 @@ class TransactionDetailsModal extends Component {
       const tx = await RPC.PROMISE('gettransaction', [
         walletItems[hoveringID].txid,
       ]);
-      this.setState({ highlightedBlockHash: tx.blockhash });
-
+      this.setState({ highlightedBlockHash: tx.blockhash,
+                      highlightedTxFee: tx.fee });
+        console.log(tx);
 
  /*
       feePromises.push(RPC.PROMISE('gettransaction', [element.txid]));
@@ -55,7 +56,7 @@ class TransactionDetailsModal extends Component {
 
       const feeAmount = await RPC.PROMISE('gettransaction', [tx.txid]);
       //let feeData = new Map([tx.time,feeAmount.fee])
-      this.default.setState({highlightedTxFee: feeAmount.fee});
+     // this.setState({highlightedTxFee: feeAmount.fee});
 
       const block = await RPC.PROMISE('getblock', [tx.blockhash]);
       this.setState({ highlightedBlockNum: block.height });
@@ -70,14 +71,15 @@ class TransactionDetailsModal extends Component {
    */
   render() {
     const { hoveringID, walletItems, settings } = this.props;
-    const { highlightedBlockNum, highlightedBlockHash } = this.state;
-
+    const { highlightedBlockNum, highlightedBlockHash, highlightedTxFee } = this.state;
+    
     if (
       hoveringID != 999999999999 &&
       !!walletItems &&
       walletItems[hoveringID]
     ) {
       const tx = walletItems[hoveringID];
+      console.log(tx.category);
       return (
         <Modal>
           <Modal.Header>Transaction Details</Modal.Header>
@@ -85,7 +87,7 @@ class TransactionDetailsModal extends Component {
             {tx.confirmations <= settings.minConfirmations && (
               <div>
                 <a>
-                  <Text id="transactions.PendingTransaction" />
+                  <Text id="transactions.PENDINGTR" />
                 </a>
               </div>
             )}
@@ -95,10 +97,10 @@ class TransactionDetailsModal extends Component {
               <span className="TXdetails">{tx.amount}</span>
             </div>
 
-            {tx.category === 'debit' && (
+            {(tx.category === 'debit'  || tx.category === 'send') && (
               <div key="modal_fee" className="detailCat">
                 <Text id="transactions.fee" />:
-                <span className="TXdetails">{+tx.fee}</span>
+                <span className="TXdetails">{highlightedTxFee}</span>
               </div>
             )}